+++ plugins/adjectives/builds/compromise-adjectives.js
@@ -334,7 +334,7 @@
           var transform = this.world.transforms.adjectives;
           this.forEach(function (doc) {
             var obj = transform(doc.text('reduced'));
-            doc.replaceWith(obj.Superlative, true, true);
+            doc.replaceWith(obj.Superlative, true);
           });
           return this;
         }
@@ -344,7 +344,7 @@
           var transform = this.world.transforms.adjectives;
           this.forEach(function (doc) {
             var obj = transform(doc.text('reduced'));
-            doc.replaceWith(obj.Comparative, true, true);
+            doc.replaceWith(obj.Comparative, true);
           });
           return this;
         }
@@ -354,7 +354,7 @@
           this.forEach(function (doc) {
             var adverb = toAdverb(doc.text('reduced'));
 
-            doc.replaceWith(adverb, true, true);
+            doc.replaceWith(adverb, true);
           });
           return this;
         }
@@ -364,7 +364,7 @@
           this.forEach(function (doc) {
             var verb = toVerb_1(doc.text('reduced'));
 
-            doc.replaceWith(verb, true, true);
+            doc.replaceWith(verb, true);
           });
           return this;
         }
@@ -374,7 +374,7 @@
           this.forEach(function (doc) {
             var noun = toNoun(doc.text('reduced'));
 
-            doc.replaceWith(noun, true, true);
+            doc.replaceWith(noun, true);
           });
           return this;
         }