+++ plugins/numbers/builds/compromise-numbers.js
@@ -990,7 +990,7 @@
         }
 
         var str = makeNumber_1(obj, val.has('#TextValue'), false);
-        val.replaceWith(str, true, true);
+        val.replaceWith(str, true);
         val.tag('Cardinal'); // turn unit into plural -> 'seven beers'
 
         _agreeUnits(agree, val, obj);
@@ -1011,7 +1011,7 @@
         }
 
         var str = makeNumber_1(obj, val.has('#TextValue'), true);
-        val.replaceWith(str, true, true);
+        val.replaceWith(str, true);
         val.tag('Ordinal'); // turn unit into singular -> 'seventh beer'
 
         var unit = _this.lookAhead('^#Plural');
@@ -1074,7 +1074,7 @@
         }
 
         var str = makeNumber_1(obj, val.has('#TextValue'), val.has('#Ordinal'));
-        val.replaceWith(str, true, true); // handle plural/singular unit
+        val.replaceWith(str, true); // handle plural/singular unit
 
         _agreeUnits(agree, val, obj);
       });
@@ -1098,7 +1098,7 @@
 
         obj.num += n;
         var str = makeNumber_1(obj, val.has('#TextValue'), val.has('#Ordinal'));
-        val.replaceWith(str, true, true); // handle plural/singular unit
+        val.replaceWith(str, true); // handle plural/singular unit
 
         _agreeUnits(agree, val, obj);
       });