+++ plugins/numbers/types/types.test.ts
@@ -1,0 +1,3 @@
+import nlp from '../'
+
+const doc = nlp('hello world').paragraphs()