+++ plugins/sentences/builds/compromise-sentences.js
@@ -148,7 +148,7 @@
       var obj = parse_1(doc);
       var vb = obj.verb.clone();
       vb = vb.verbs().toNegative();
-      obj.verb.replaceWith(vb, false, true);
+      obj.verb.replaceWith(vb, false);
     });
     return this;
   };
@@ -160,7 +160,7 @@
       var obj = parse_1(doc);
       var vb = obj.verb.clone();
       vb = vb.verbs().toPositive();
-      obj.verb.replaceWith(vb, false, true);
+      obj.verb.replaceWith(vb, false);
     });
     return this;
   };
@@ -234,7 +234,7 @@
       var obj = parse_1(doc);
       var vb = obj.verb.clone();
       vb = vb.verbs().toPastTense();
-      obj.verb.replaceWith(vb, false, true); // trailing gerund/future/present are okay, but 'walked and eats' is not
+      obj.verb.replaceWith(vb, false); // trailing gerund/future/present are okay, but 'walked and eats' is not
 
       if (obj.object && obj.object.found && obj.object.has('#PresentTense')) {
         var verbs = obj.object.verbs();
@@ -264,7 +264,7 @@
         vb = vb.verbs().toPresentTense();
       }
 
-      obj.verb.replaceWith(vb, false, true); // future is okay, but 'walks and ate' -> 'walks and eats'
+      obj.verb.replaceWith(vb, false); // future is okay, but 'walks and ate' -> 'walks and eats'
 
       if (obj.object && obj.object.found && obj.object.has('#PastTense')) {
         var verbs = obj.object.verbs();
@@ -281,7 +281,7 @@
       var obj = parse_1(doc);
       var vb = obj.verb.clone();
       vb = vb.verbs().toFutureTense();
-      obj.verb.replaceWith(vb, false, true); //Present is okay, but 'will walk and ate' -> 'will walk and eat'
+      obj.verb.replaceWith(vb, false); //Present is okay, but 'will walk and ate' -> 'will walk and eat'
 
       if (obj.object && obj.object.found && obj.object.has('(#PastTense|#PresentTense)')) {
         var verbs = obj.object.verbs();