+++ src/02-tagger/03-contractions/02-simple.js
@@ -1,4 +1,4 @@
-const contraction = /([a-z\u00C0-\u00FF]+)'([a-z]{1,2})$/i
+const contraction = /([a-z\u00C0-\u00FF]+)[\u0027\u0060\u00B4\u2018\u2019\u201A\u201B\u2032\u2035\u2039\u203A]([a-z]{1,2})$/i
 
 //these ones don't seem to be ambiguous
 const easy = {