+++ src/02-tagger/03-contractions/03-irregulars.js
@@ -10,9 +10,9 @@ const irregulars = {
   whered: ['where', 'did'],
   "when'd": ['when', 'did'],
   whend: ['when', 'did'],
-  "how'd": ['how', 'did'],
+  // "how'd": ['how', 'did'], //'how would?'
+  // "what'd": ['what', 'did'], //'what would?'
   howd: ['how', 'did'],
-  "what'd": ['what', 'did'],
   whatd: ['what', 'did'],
   // "let's": ['let', 'us'], //too weird
 
@@ -29,27 +29,8 @@ const irregulars = {
   '@': ['at'],
 }
 
-// either 'is not' or 'are not'
-const doAint = function(term, phrase) {
-  let terms = phrase.cache.terms || phrase.terms()
-  let index = terms.indexOf(term)
-  let before = terms.slice(0, index)
-  //look for the preceding noun
-  let noun = before.find(t => {
-    return t.tags.Noun
-  })
-  if (noun && noun.tags.Plural) {
-    return ['are', 'not']
-  }
-  return ['is', 'not']
-}
-
 //
-const checkIrregulars = function(term, phrase) {
-  //this word needs it's own logic:
-  if (term.clean === `ain't` || term.clean === 'aint') {
-    return doAint(term, phrase)
-  }
+const checkIrregulars = function(term) {
   //check white-list
   if (irregulars.hasOwnProperty(term.clean)) {
     return irregulars[term.clean]