+++ tests/contractions/contraction.test.js
@@ -1,6 +1,52 @@
 const test = require('tape')
 const nlp = require('../_lib')
 
+test('tricky contractions', function(t) {
+  let doc = nlp(`I’m `)
+  doc.contractions().expand()
+  t.equal(doc.text(), 'I am ', '')
+
+  doc = nlp(` i can’t `)
+  doc.contractions().expand()
+  t.equal(doc.text(), ' i can not ', '')
+
+  doc = nlp(`spencer’s clean`)
+  doc.contractions().expand()
+  t.equal(doc.text(), 'spencer is clean', '')
+
+  doc = nlp(`wouldn’t be good`)
+  doc.contractions().expand()
+  t.equal(doc.text(), 'would not be good', '')
+
+  doc = nlp(`what’d you see`)
+  doc.contractions().expand()
+  t.equal(doc.text(), 'what did you see', '')
+
+  doc = nlp(`spencer’d go see`)
+  doc.contractions().expand()
+  t.equal(doc.text(), 'spencer would go see', '')
+
+  t.end()
+})
+
+test('contractions v possessive', function(t) {
+  let str = nlp("spencer's not cool")
+    .normalize({ contractions: true })
+    .text()
+  t.equal(str, 'spencer is not cool', 'adj contraction')
+
+  str = nlp("spencer's walk")
+    .normalize({ contractions: true })
+    .text()
+  t.equal(str, "spencer's walk", 'noun not contraction')
+
+  str = nlp("spencer's runs")
+    .normalize({ contractions: true })
+    .text()
+  t.equal(str, "spencer's runs", 'present-tense not contraction')
+  t.end()
+})
+
 test('match contractions/possessives', function(t) {
   let doc = nlp(`i think it's spencer's`)
   t.equal(doc.has('it'), true, 'has it')